+++ src/encoder.ts
@@ -1,14 +1,14 @@
-import {ColorSpace, ColorBook} from './types';
+import {ColorModel, ColorBook} from './types';
 import validate from './validator';
 import * as Chunk from './chunk';
 
-const ColorSpaceToId: Record<ColorSpace, number> = {
+const ColorModelToId: Record<ColorModel, number> = {
   RGB: 0,
   CMYK: 2,
   Lab: 7,
 };
 
-const ColorSpaceComponents: Record<ColorSpace, number> = {
+const ColorModelComponents: Record<ColorModel, number> = {
   RGB: 3,
   CMYK: 4,
   Lab: 3,
@@ -29,20 +29,20 @@ export function* encodeAcb(book: ColorBook) {
   yield Chunk.fromUInt16BE(book.id);
   yield Chunk.fromString(book.title);
   yield Chunk.fromString(book.colorNamePrefix);
-  yield Chunk.fromString(book.colorNameSuffix);
+  yield Chunk.fromString(book.colorNamePostfix);
   yield Chunk.fromString(book.description);
   yield Chunk.fromUInt16BE(book.colors.length);
   yield Chunk.fromUInt16BE(book.pageSize);
-  yield Chunk.fromUInt16BE(book.pageMidPoint);
+  yield Chunk.fromUInt16BE(book.pageKey);
 
-  const colorSpaceId = ColorSpaceToId[book.colorSpace];
-  const expectedComponents = ColorSpaceComponents[book.colorSpace];
+  const colorModelId = ColorModelToId[book.colorModel];
+  const expectedComponents = ColorModelComponents[book.colorModel];
 
-  if (isNaN(colorSpaceId)) {
-    throw new Error(`Unknown color space: ${book.colorSpace}`);
+  if (isNaN(colorModelId)) {
+    throw new Error(`Unknown color space: ${book.colorModel}`);
   }
 
-  yield Chunk.fromUInt16BE(colorSpaceId);
+  yield Chunk.fromUInt16BE(colorModelId);
 
   for (let color of book.colors) {
     if (color.code.length !== 6) {
@@ -55,7 +55,7 @@ export function* encodeAcb(book: ColorBook) {
 
     yield Chunk.fromString(color.name);
     yield Chunk.fromAscii(color.code);
-    yield Chunk.fromComponents(color.components, book.colorSpace);
+    yield Chunk.fromComponents(color.components, book.colorModel);
   }
 
   yield Chunk.fromAscii(book.isSpot ? 'spflspot' : 'spflproc');