+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Renamed properties with more accurate terms: colorSpace -> colorModel,
+  colorNameSuffix -> colorNamePostfix, pageMidPoint -> pageKey
+
 ## [0.1.0] - 2021-08-21
 
 ### Added